requirements.txt
# File: requirements.txt
# Python package dependencies for the Iroh Home Management System

# HTTP client for async operations
httpx>=0.24.0

# YAML configuration support
PyYAML>=6.0.1

# WebSocket client for phone API
websockets>=11.0.3

# Audio processing
numpy>=1.24.0

# Async support
aiofiles>=23.1.0
asyncio>=3.4.3

# Type hints
typing-extensions>=4.7.1

# Logging
python-json-logger>=2.0.7

# Home Assistant integration
aiohttp>=3.8.5

# Development tools
black>=23.7.0
pylint>=2.17.5
mypy>=1.5.1
